pyproject.toml
[project]
name = "detect-jailbreak"
version = "0.1.0"
description = "A prompt-injection and jailbreak detector for LLMs."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"},
    {name = "Joseph Catrambone", email="joseph@guardrailsai.com"},
    {name = "Brian Lai", email="brian@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.9.0"
dependencies = [
    "guardrails-ai>=0.5.10"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
] 

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "tests"
]

[tool.pyright]
include = ["validator"]